refactor: Extracted methods from ast-jsx-utils and types that are only used in react-base-plugin
#244
