+++ packages/teleport-plugin-react-base-component/src/constants.ts
@@ -1,0 +1,13 @@
+export const ERROR_LOG_NAME = `react-base-component`
+
+export const REACT_LIBRARY_DEPENDENCY = {
+  type: 'library',
+  path: 'react',
+  version: '16.8.3',
+}
+
+export const DEFAULT_COMPONENT_CHUNK_NAME = 'react-component'
+
+export const DEFAULT_EXPORT_CHUNK_NAME = 'export'
+
+export const DEFAULT_IMPORT_CHUNK_NAME = 'import-local'
