+++ packages/teleport-plugin-react-base-component/src/utils.ts
@@ -2,206 +2,32 @@ import * as types from '@babel/types'
 
 import { convertValueToLiteral } from '@teleporthq/teleport-shared/lib/utils/ast-js-utils'
 import {
-  addChildJSXTag,
-  addChildJSXText,
   addAttributeToJSXTag,
   addDynamicAttributeOnTag,
-  createConditionalJSXExpression,
-  generateASTDefinitionForJSXTag,
 } from '@teleporthq/teleport-shared/lib/utils/ast-jsx-utils'
-
 import { capitalize } from '@teleporthq/teleport-shared/lib/utils/string-utils'
+
 import {
-  UIDLElementNode,
   UIDLPropDefinition,
-  UIDLRepeatNode,
-  UIDLConditionalNode,
-  UIDLConditionalExpression,
   UIDLAttributeValue,
   UIDLDynamicReference,
   ComponentDependency,
   UIDLStateDefinition,
   EventHandlerStatement,
-  UIDLSlotNode,
   StateIdentifier,
-  NodeSyntaxGenerator,
   AttributeAssignCodeMod,
   ConditionalIdentifier,
+  UIDLConditionalExpression,
 } from '@teleporthq/teleport-types'
 
-import { ERROR_LOG_NAME } from '.'
-
-interface ReactComponentAccumulators {
-  propDefinitions: Record<string, UIDLPropDefinition>
-  stateIdentifiers: Record<string, StateIdentifier>
-  nodesLookup: Record<string, types.JSXElement>
-  dependencies: Record<string, ComponentDependency>
-}
-
-export const generateElementNode = (
-  node: UIDLElementNode,
-  accumulators: ReactComponentAccumulators
-) => {
-  const { dependencies, stateIdentifiers, propDefinitions, nodesLookup } = accumulators
-  const { elementType, children, key, attrs, dependency, events } = node.content
-  const elementTag = generateASTDefinitionForJSXTag(elementType)
-
-  if (attrs) {
-    Object.keys(attrs).forEach((attrKey) => {
-      addAttributeToNode(elementTag, attrKey, attrs[attrKey])
-    })
-  }
-
-  if (dependency) {
-    // Make a copy to avoid reference leaking
-    dependencies[elementType] = { ...dependency }
-  }
-
-  if (events) {
-    Object.keys(events).forEach((eventKey) => {
-      addEventHandlerToTag(
-        elementTag,
-        eventKey,
-        events[eventKey],
-        stateIdentifiers,
-        propDefinitions
-      )
-    })
-  }
-
-  if (children) {
-    children.forEach((child) => {
-      const childTag = generateNodeSyntax(child, accumulators)
-
-      if (typeof childTag === 'string') {
-        addChildJSXText(elementTag, childTag)
-      } else if (childTag.type === 'JSXExpressionContainer') {
-        addChildJSXTag(elementTag, childTag)
-      } else {
-        addChildJSXTag(elementTag, types.jsxExpressionContainer(childTag))
-      }
-    })
-  }
-
-  nodesLookup[key] = elementTag
-  return elementTag
-}
-
-export const generateRepeatNode = (
-  node: UIDLRepeatNode,
-  accumulators: ReactComponentAccumulators
-) => {
-  const { node: repeatContent, dataSource, meta } = node.content
-
-  const contentAST = generateNodeSyntax(repeatContent, accumulators)
-
-  if (typeof contentAST === 'string' || (contentAST as types.JSXExpressionContainer).expression) {
-    throw new Error(
-      `${ERROR_LOG_NAME} generateRepeatNode found a repeat node that specified invalid content ${JSON.stringify(
-        contentAST,
-        null,
-        2
-      )}`
-    )
-  }
-
-  const repeatAST = makeRepeatStructureWithMap(dataSource, contentAST as types.JSXElement, meta)
-  return repeatAST
-}
-
-export const generateConditionalNode = (
-  node: UIDLConditionalNode,
-  accumulators: ReactComponentAccumulators
-) => {
-  const { reference, value } = node.content
-  const conditionIdentifier = createConditionIdentifier(reference, accumulators)
-
-  const subTree = generateNodeSyntax(node.content.node, accumulators)
-
-  const condition: UIDLConditionalExpression =
-    value !== undefined && value !== null
-      ? { conditions: [{ operand: value, operation: '===' }] }
-      : node.content.condition
-
-  return createConditionalJSXExpression(subTree, condition, conditionIdentifier)
-}
-
-export const generateSlotNode = (
-  node: UIDLSlotNode,
-  accumulators: ReactComponentAccumulators,
-  t = types
-) => {
-  // TODO: Handle multiple slots with props['slot-name']
-  const childrenProp: UIDLDynamicReference = {
-    type: 'dynamic',
-    content: {
-      referenceType: 'prop',
-      id: 'children',
-    },
-  }
-
-  const childrenExpression = makeDynamicValueExpression(childrenProp)
-
-  if (node.content.fallback) {
-    const fallbackContent = generateNodeSyntax(node.content.fallback, accumulators)
-    let expression: types.Expression
-
-    if (typeof fallbackContent === 'string') {
-      expression = t.stringLiteral(fallbackContent)
-    } else if ((fallbackContent as types.JSXExpressionContainer).expression) {
-      expression = (fallbackContent as types.JSXExpressionContainer).expression as types.Expression
-    } else {
-      expression = fallbackContent as types.JSXElement
-    }
-
-    // props.children with fallback
-    return t.jsxExpressionContainer(t.logicalExpression('||', childrenExpression, expression))
-  }
-
-  return t.jsxExpressionContainer(childrenExpression)
-}
-
-type GenerateNodeSyntaxReturnValue =
-  | string
-  | types.JSXExpressionContainer
-  | types.JSXElement
-  | types.LogicalExpression
-  | types.Identifier
-  | types.MemberExpression
-
-export const generateNodeSyntax: NodeSyntaxGenerator<
+import { ERROR_LOG_NAME } from './constants'
+import {
+  GenerateNodeSyntaxReturnValue,
   ReactComponentAccumulators,
-  GenerateNodeSyntaxReturnValue
-> = (node, accumulators) => {
-  switch (node.type) {
-    case 'static':
-      return node.content.toString()
-
-    case 'dynamic':
-      return makeDynamicValueExpression(node)
-
-    case 'element':
-      return generateElementNode(node, accumulators)
-
-    case 'repeat':
-      return generateRepeatNode(node, accumulators)
-
-    case 'conditional':
-      return generateConditionalNode(node, accumulators)
-
-    case 'slot':
-      return generateSlotNode(node, accumulators)
-
-    default:
-      throw new Error(
-        `${ERROR_LOG_NAME} generateNodeSyntax encountered a node of unsupported type: ${JSON.stringify(
-          node,
-          null,
-          2
-        )}`
-      )
-  }
-}
+  BinaryOperator,
+  UnaryOperation,
+  ContentType,
+} from './types'
 
 export const createStateIdentifiers = (
   stateDefinitions: Record<string, UIDLStateDefinition>,
@@ -266,22 +92,9 @@ export const makePureComponent = (
   return component
 }
 
-const generateReturnExpressionSyntax = (
-  stateIdentifiers: Record<string, StateIdentifier>,
-  jsxTagTree: types.JSXElement,
-  t = types
-) => {
-  const returnStatement = t.returnStatement(jsxTagTree)
-  const stateHooks = Object.keys(stateIdentifiers).map((stateKey) =>
-    makeStateHookAST(stateIdentifiers[stateKey])
-  )
-
-  return t.blockStatement([...stateHooks, returnStatement] || [])
-}
-
 // Adds all the event handlers and all the instructions for each event handler
 // in case there is more than one specified in the UIDL
-const addEventHandlerToTag = (
+export const addEventHandlerToTag = (
   tag: types.JSXElement,
   eventKey: string,
   eventHandlerStatements: EventHandlerStatement[],
@@ -322,6 +135,225 @@ const addEventHandlerToTag = (
   tag.openingElement.attributes.push(
     t.jsxAttribute(t.jsxIdentifier(eventKey), t.jsxExpressionContainer(expressionContent))
   )
+}
+
+export const makeRepeatStructureWithMap = (
+  dataSource: UIDLAttributeValue,
+  content: types.JSXElement,
+  meta: Record<string, any> = {},
+  t = types
+) => {
+  const iteratorName = meta.iteratorName || 'item'
+  const keyIdentifier = meta.useIndex ? 'index' : iteratorName
+
+  const source = getSourceIdentifier(dataSource)
+
+  const dynamicLocalReference: UIDLDynamicReference = {
+    type: 'dynamic',
+    content: {
+      referenceType: 'local',
+      id: keyIdentifier,
+    },
+  }
+  addAttributeToNode(content, 'key', dynamicLocalReference)
+
+  const arrowFunctionArguments = [t.identifier(iteratorName)]
+  if (meta.useIndex) {
+    arrowFunctionArguments.push(t.identifier('index'))
+  }
+
+  return t.jsxExpressionContainer(
+    t.callExpression(t.memberExpression(source, t.identifier('map')), [
+      t.arrowFunctionExpression(arrowFunctionArguments, content),
+    ])
+  )
+}
+
+export const makeDynamicValueExpression = (identifier: UIDLDynamicReference, t = types) => {
+  const prefix = getReactVarNameForDynamicReference(identifier)
+  return prefix === ''
+    ? t.identifier(identifier.content.id)
+    : t.memberExpression(t.identifier(prefix), t.identifier(identifier.content.id))
+}
+
+// Prepares an identifier (from props or state) to be used as a conditional rendering identifier
+// Assumes the type from the corresponding props/state definitions
+export const createConditionIdentifier = (
+  dynamicReference: UIDLDynamicReference,
+  accumulators: ReactComponentAccumulators
+): ConditionalIdentifier => {
+  const { id, referenceType } = dynamicReference.content
+
+  switch (referenceType) {
+    case 'prop':
+      return {
+        key: id,
+        type: accumulators.propDefinitions[id].type,
+        prefix: 'props',
+      }
+    case 'state':
+      return {
+        key: id,
+        type: accumulators.stateIdentifiers[id].type,
+      }
+    default:
+      throw new Error(
+        `${ERROR_LOG_NAME} createConditionIdentifier encountered an invalid reference type: ${JSON.stringify(
+          dynamicReference,
+          null,
+          2
+        )}`
+      )
+  }
+}
+
+/**
+ * @param tag the ref to the AST tag under construction
+ * @param attributeKey the key of the attribute that should be added on the current AST node
+ * @param attributeValue the value(string, number, bool) of the attribute that should be added on the current AST node
+ */
+export const addAttributeToNode: AttributeAssignCodeMod<types.JSXElement> = (
+  tag,
+  attributeKey,
+  attributeValue
+) => {
+  switch (attributeValue.type) {
+    case 'dynamic':
+      const {
+        content: { id },
+      } = attributeValue
+      const prefix = getReactVarNameForDynamicReference(attributeValue)
+      addDynamicAttributeOnTag(tag, attributeKey, id, prefix)
+      return
+    case 'static':
+      const { content } = attributeValue
+      addAttributeToJSXTag(tag, { name: attributeKey, value: content })
+      return
+    default:
+      throw new Error(
+        `${ERROR_LOG_NAME} addAttributeToNode could not generate code for assignment of type ${JSON.stringify(
+          attributeValue
+        )}`
+      )
+  }
+}
+
+export const createConditionalJSXExpression = (
+  content: ContentType,
+  conditionalExpression: UIDLConditionalExpression,
+  conditionalIdentifier: ConditionalIdentifier,
+  t = types
+) => {
+  let contentNode: types.Expression
+
+  if (typeof content === 'string') {
+    contentNode = t.stringLiteral(content)
+  } else if (content.type === 'JSXExpressionContainer') {
+    contentNode = content.expression as types.Expression
+  } else {
+    contentNode = content
+  }
+
+  let binaryExpression:
+    | types.LogicalExpression
+    | types.BinaryExpression
+    | types.UnaryExpression
+    | types.Identifier
+    | types.MemberExpression
+
+  // When the stateValue is an object we will compute a logical/binary expression on the left side
+  const { conditions, matchingCriteria } = conditionalExpression
+  const binaryExpressions = conditions.map((condition) =>
+    createBinaryExpression(condition, conditionalIdentifier)
+  )
+
+  if (binaryExpressions.length === 1) {
+    binaryExpression = binaryExpressions[0]
+  } else {
+    // the first two binary expressions are put together as a logical expression
+    const [firstExp, secondExp] = binaryExpressions
+    const operation = matchingCriteria === 'all' ? '&&' : '||'
+    let expression: types.LogicalExpression = t.logicalExpression(operation, firstExp, secondExp)
+
+    // accumulate the rest of the expressions to the logical expression
+    for (let index = 2; index < binaryExpressions.length; index++) {
+      expression = t.logicalExpression(operation, expression, binaryExpressions[index])
+    }
+
+    binaryExpression = expression
+  }
+
+  return t.logicalExpression('&&', binaryExpression, contentNode)
+}
+
+export const createBinaryExpression = (
+  condition: {
+    operation: string
+    operand?: string | number | boolean
+  },
+  conditionalIdentifier: ConditionalIdentifier,
+  t = types
+) => {
+  const { operand, operation } = condition
+  const identifier = conditionalIdentifier.prefix
+    ? t.memberExpression(
+        t.identifier(conditionalIdentifier.prefix),
+        t.identifier(conditionalIdentifier.key)
+      )
+    : t.identifier(conditionalIdentifier.key)
+
+  if (operation === '===') {
+    if (operand === true) {
+      return identifier
+    }
+
+    if (operand === false) {
+      return t.unaryExpression('!', identifier)
+    }
+  }
+
+  if (operand !== undefined) {
+    const stateValueIdentifier = convertValueToLiteral(operand, conditionalIdentifier.type)
+
+    return t.binaryExpression(convertToBinaryOperator(operation), identifier, stateValueIdentifier)
+  } else {
+    return operation ? t.unaryExpression(convertToUnaryOperator(operation), identifier) : identifier
+  }
+}
+
+/**
+ * Because of the restrictions of the AST Types we need to have a clear subset of binary operators we can use
+ * @param operation - the operation defined in the UIDL for the current state branch
+ */
+const convertToBinaryOperator = (operation: string): BinaryOperator => {
+  const allowedOperations = ['===', '!==', '>=', '<=', '>', '<']
+  if (allowedOperations.includes(operation)) {
+    return operation as BinaryOperator
+  } else {
+    return '==='
+  }
+}
+
+const convertToUnaryOperator = (operation: string): UnaryOperation => {
+  const allowedOperations = ['!']
+  if (allowedOperations.includes(operation)) {
+    return operation as UnaryOperation
+  } else {
+    return '!'
+  }
+}
+
+const generateReturnExpressionSyntax = (
+  stateIdentifiers: Record<string, StateIdentifier>,
+  jsxTagTree: types.JSXElement,
+  t = types
+) => {
+  const returnStatement = t.returnStatement(jsxTagTree)
+  const stateHooks = Object.keys(stateIdentifiers).map((stateKey) =>
+    makeStateHookAST(stateIdentifiers[stateKey])
+  )
+
+  return t.blockStatement([...stateHooks, returnStatement] || [])
 }
 
 const createPropCallStatement = (
@@ -391,76 +423,6 @@ const makeStateHookAST = (stateIdentifier: StateIdentifier, t = types) => {
   ])
 }
 
-const makeRepeatStructureWithMap = (
-  dataSource: UIDLAttributeValue,
-  content: types.JSXElement,
-  meta: Record<string, any> = {},
-  t = types
-) => {
-  const iteratorName = meta.iteratorName || 'item'
-  const keyIdentifier = meta.useIndex ? 'index' : iteratorName
-
-  const source = getSourceIdentifier(dataSource)
-
-  const dynamicLocalReference: UIDLDynamicReference = {
-    type: 'dynamic',
-    content: {
-      referenceType: 'local',
-      id: keyIdentifier,
-    },
-  }
-  addAttributeToNode(content, 'key', dynamicLocalReference)
-
-  const arrowFunctionArguments = [t.identifier(iteratorName)]
-  if (meta.useIndex) {
-    arrowFunctionArguments.push(t.identifier('index'))
-  }
-
-  return t.jsxExpressionContainer(
-    t.callExpression(t.memberExpression(source, t.identifier('map')), [
-      t.arrowFunctionExpression(arrowFunctionArguments, content),
-    ])
-  )
-}
-
-const makeDynamicValueExpression = (identifier: UIDLDynamicReference, t = types) => {
-  const prefix = getReactVarNameForDynamicReference(identifier)
-  return prefix === ''
-    ? t.identifier(identifier.content.id)
-    : t.memberExpression(t.identifier(prefix), t.identifier(identifier.content.id))
-}
-
-// Prepares an identifier (from props or state) to be used as a conditional rendering identifier
-// Assumes the type from the corresponding props/state definitions
-const createConditionIdentifier = (
-  dynamicReference: UIDLDynamicReference,
-  accumulators: ReactComponentAccumulators
-): ConditionalIdentifier => {
-  const { id, referenceType } = dynamicReference.content
-
-  switch (referenceType) {
-    case 'prop':
-      return {
-        key: id,
-        type: accumulators.propDefinitions[id].type,
-        prefix: 'props',
-      }
-    case 'state':
-      return {
-        key: id,
-        type: accumulators.stateIdentifiers[id].type,
-      }
-    default:
-      throw new Error(
-        `${ERROR_LOG_NAME} createConditionIdentifier encountered an invalid reference type: ${JSON.stringify(
-          dynamicReference,
-          null,
-          2
-        )}`
-      )
-  }
-}
-
 const getSourceIdentifier = (dataSource: UIDLAttributeValue, t = types) => {
   switch (dataSource.type) {
     case 'static':
@@ -482,35 +444,4 @@ const getReactVarNameForDynamicReference = (dynamicReference: UIDLDynamicReferen
     static: '',
     local: '',
   }[dynamicReference.content.referenceType]
-}
-
-/**
- * @param tag the ref to the AST tag under construction
- * @param attributeKey the key of the attribute that should be added on the current AST node
- * @param attributeValue the value(string, number, bool) of the attribute that should be added on the current AST node
- */
-const addAttributeToNode: AttributeAssignCodeMod<types.JSXElement> = (
-  tag,
-  attributeKey,
-  attributeValue
-) => {
-  switch (attributeValue.type) {
-    case 'dynamic':
-      const {
-        content: { id },
-      } = attributeValue
-      const prefix = getReactVarNameForDynamicReference(attributeValue)
-      addDynamicAttributeOnTag(tag, attributeKey, id, prefix)
-      return
-    case 'static':
-      const { content } = attributeValue
-      addAttributeToJSXTag(tag, { name: attributeKey, value: content })
-      return
-    default:
-      throw new Error(
-        `${ERROR_LOG_NAME} addAttributeToNode could not generate code for assignment of type ${JSON.stringify(
-          attributeValue
-        )}`
-      )
-  }
 }
